pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bitbank-bot"
version = "0.1.0"
description = "Bitbank cryptocurrency auto-trading bot with trend-following strategy"
requires-python = ">=3.9"
dependencies = [
    "ccxt>=4.0",
    "pandas>=2.0",
    "ta>=0.10",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "requests>=2.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "flake8>=6.0",
]

[project.scripts]
bitbank-bot = "bitbank_bot.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
